requirements.txt
transformers
llama_index==0.9.46
python-dotenv==1.0.1

